requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
traci>=1.18.0
sumolib>=1.18.0
gymnasium>=0.28.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
pytest>=7.3.0
